# src/lib/swagger-spec-generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Parcours Moniteur",
    "version": "0.1.0",
    "description": "API pour gérer les parcours et itinéraires"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Vérifier la santé de l'API et la connexion à la BDD",
        "description": "Teste la connexion à Supabase et retourne le statut",
        "tags": [
          "Health"
        ],
        "responses": {
          "200": {
            "description": "API et BDD fonctionnent correctement",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "database": {
                      "type": "string",
                      "example": "connected"
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur de connexion à la BDD"
          }
        }
      }
    },
    "/api/pois/{id}": {
      "get": {
        "summary": "Get a specific POI by ID",
        "description": "Retrieve a Point of Interest by its UUID",
        "tags": [
          "POIs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "POI unique identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "POI found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "name": {
                          "type": "string"
                        },
                        "latitude": {
                          "type": "number"
                        },
                        "longitude": {
                          "type": "number"
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updated_at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid POI ID format"
          },
          "404": {
            "description": "POI not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/pois": {
      "get": {
        "summary": "Get all POIs",
        "description": "Retrieve all Points of Interest",
        "tags": [
          "POIs"
        ],
        "responses": {
          "200": {
            "description": "List of all POIs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "name": {
                        "type": "string"
                      },
                      "latitude": {
                        "type": "number"
                      },
                      "longitude": {
                        "type": "number"
                      },
                      "created_at": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updated_at": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "tags": []
}